requirements.txt
asgiref==3.7.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.0.2
django-froala-editor==4.5.2
django-widget-tweaks==1.5.0
idna==3.10
pillow==11.2.1
requests==2.32.3
sqlparse==0.4.4
tailwind==3.1.5b0
typing_extensions==4.9.0
urllib3==2.4.0
